shop-agent-app/requirements.txt
fastmcp==2.10.4
google-adk==1.5.0
google-cloud-retail==2.3.0
python-dotenv==1.1.1
google-auth==2.40.3
